requirements.txt
fastapi
torch
torchvision
torchaudio
pillow
diffusers["torch"]
transformers
